feat: refactor ApiKeyDialog and useAuthSearchTool for improved API key management
